Backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ── AI / LLM ──────────────────────────────────────────────────────────────────
openai==1.40.0

# ── Config & Validation ───────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.8.2

# ── HTTP Client ──────────────────────────────────────────────────────────────
httpx==0.27.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
